+++ app/jobs/application_job.rb
@@ -17,43 +17,29 @@ class ApplicationJob < ActiveJob::Base
 
   # info for health_check
   # content of @@last_job_warnings should be added with leading line feed
-  def self.last_job_warnings(job_class)
+  # !! If jobs are not still alive this should be handled by restart of Docker container after some time of unhealthy container !!
+  def self.last_job_warnings
     retval = ''                                                                 # default for :ok
-    job_info = @@last_job_warnings[job_class.name]
 
-    retval << job_info[:message] if job_info && !job_info[:message].nil?
+    # Number of cycles to wait before assuming job is not active no more and warning is issued
+    wait_factor = 10
 
-    if job_info && Time.now > job_info[:last_execution] + job_info[:cycle_seconds]
-      retval << "\n" if retval != ''
-      retval << "Last execution of Job #{job_class} (#{job_info[:last_execution]}) is older than now - #{job_info[:cycle_seconds]} seconds "
-      retval << "(#{job_info[:cycle_seconds]/60} minutes, #{job_info[:cycle_seconds]/3600} hours, #{job_info[:cycle_seconds]/(3600*24)} days ), "
-      retval << "but should occure every #{job_info[:cycle_seconds]} seconds! Please check for sufficient memory and restart MOVEX CDC to fix the issue."
+    @@last_job_warnings.each do |job_class, job_info|
+      retval << job_info[:message] if job_info && !job_info[:message].nil?
+
+      # If job_info is nil, it means that the job has not been enqueued yet
+      if job_info && Time.now > job_info[:last_execution] + job_info[:cycle_seconds] * wait_factor
+        message = "Last execution of Job #{job_class} (#{job_info[:last_execution]}) is older than now - #{job_info[:cycle_seconds]} seconds "
+        message << "(#{job_info[:cycle_seconds]/60} minutes, #{job_info[:cycle_seconds]/3600} hours, #{job_info[:cycle_seconds]/(3600*24)} days ), "
+        message << "but should occure every #{job_info[:cycle_seconds]} seconds! Please check for sufficient memory and restart MOVEX CDC to fix the issue."
+        Rails.logger.warn('ApplicationJob.last_job_warnings'){ message }
+
+        retval << "\n" if retval != ''
+        retval << message
+      end
     end
     retval
   end
-
-  # Sometimes at OutOfMemory conditions jobs are not restarted and remain inactive for the future
-  # Housekeeping executed by Docker container can repair this seldom scenario
-  def self.ensure_job_rescheduling
-    if @@reschedule_mutex.locked?
-      Rails.logger.warn('ApplicationJob.ensure_job_restarts'){"Mutex is locked by another thread. Not waiting."}
-      return                                                                    # do not wait for mutex in health_check
-    end
-    @@reschedule_mutex.synchronize do
-      wait_factor = 2                                                             # reschedule jobs if not executed within twice the expected cycle
-      @@last_job_warnings.each do |job_class_name, value|
-        if Time.now > value[:last_execution] + value[:cycle_seconds] * wait_factor    # Wait twice the cycle before assuming job as not active no more
-          SystemValidationJob.set(wait: SystemValidationJob::CYCLE.seconds).perform_later unless Rails.env.test?  # Ensure next execution independent from following operations
-          Rails.logger.warn('ApplicationJob.ensure_job_restarts'){ "Job '#{job_class_name}' has not been executed for #{wait_factor} * cycle_seconds (#{value[:cycle_seconds]})!"}
-          Rails.logger.warn('ApplicationJob.ensure_job_restarts'){ "Last execution time for job '#{job_class_name}' was #{value[:last_execution]}."}
-          Rails.logger.warn('ApplicationJob.ensure_job_restarts'){ "This may happen randomly if application runs out of memory."}
-          Rails.logger.warn('ApplicationJob.ensure_job_restarts'){ "Rescheduling job '#{job_class_name}' for now + #{value[:cycle_seconds]} seconds."}
-          job_class_name.constantize.set(wait: value[:cycle_seconds].seconds).perform_later # unless Rails.env.test?
-        end
-      end
-    end
-  end
-
 
   def self.job_infos
     @@last_job_warnings